Add conditional plugin loading in the settings.gradle.kts file
Travis is not able to build the project as the Gradle plugin
1.1.0-SNAPSHOT is missing in all the repository. This is preventing the
project from even building the plugin and publishing it to Maven Local.

To avoid this, I'm adding a conditional loading of the Gradle modules
that will load the `samples` only if we the plugin is found in any
repository.