Specify Java source and target compatibility version
Add the directive to specify the version of Java to use
to compile the Java files. Previously the version was not specified
so the local JDK was used.

This caused a problem with version 1.1.0 of the plugin that got compiled
with Java 11 and was incompatible for users on Java 8.

Fixes #38